server: Allow scans without filename
The generic multipart/form-data parser in sanic distinguishes file
uploads from text fields using the presence of the filename or filename*
attributes in the Content-Disposition header. Is has no other chance
since it wants to be generic. This essentially makes it impossible to
submit a sample to our API without specifying the filename attribute
because the generic parser would then treat the field as a text field
and try to decode it using the form charset (or utf-8 as fallback, and
croak on it most of the time).

We could require clients to specify an empty filename. Apart from being
a workaround it muddies the distinction between a sample having no file
name and an empty file name declared in the original mail. While this
seems a pedantic distinction, the exploration of how to remedy this
issue was an eye-opener in terms of adjacent problems such as encoding
of file names as well as alternative form data formats.

So this change pulls the multipart/form-data parser of sanic into our
server. There we can modify it to:

- accept only one field 'file',
- embed the knowledge that this field only accepts binary data and
- therefore doesn't need a filename attribute to heuristically detect
  that as well as
- deny accepting application/x-www-form-urlencoded bodies because
  they're generally agreed not to be suited for uploading binary data
  and we never really supported it before because we relied on sanic to
  extract files from the form data and with urlencoded it cannot do
  that. This doesn't preclude us from adding support for that format in
  the future. It just clarifies an implicit assumption we made about form
  data before.

We also add a number of explanatory comments regarding the encoding
pitfalls within the scheme. The only way to transfer file names without
corruption currently turns out to be RFC2231 encoding in a 'filename*'
parameter (RFC2388).