amavis: Improve filename encoding handling
We noticed that LWP::UserAgent of libwww-perl breaks filename encoding
during POST to the Peekaboo REST API because it converts the internal
string representation of Perl into latin1 on output.

This change steps around that issue by encoding the filename using
RFC2231 (i.e. RFC2388) in the multipart/form-data structure as this has
turned out to be the only way to get file names into Peekaboo without
corruption for the time being. This also changes plugin behaviour to no
longer submit the local file name (p00x) as a fallback if no declared
name is available.

During research into the issue it became clear, that extracting a
useable filename from AMaViS is even harder, mostly because encoding
standards in MIME have an even longer history than in HTTP but also
because the interface AMaViS provides ($part->name_declared array) is
lossy in terms of context. Therefore we switch from decoding the name
ourselves to adding a number of heuristics to extract the best possible
filename from AMaViS and add a lot of explanatory documentation for the
next time this will invariably break.

As a showcase we make our scan_emailed_file client more compliant by
adding RFC2231 encoding there as well.

Finally while working on this, it became clear that the plugin was still
overly complex and could be simplified. We would loop over various part
attributes, extract them in to a hash (once meant for serialisation into
JSON for direct submit to Peekaboo) and then reference them from there.
We can and do just drop that unnecessary step of extraction into the
hash.