daemon: Detect PID re-use
Detect if the PID file contains our own pid and treat it as stale if so.
In case of misconfiguration we still issue the warning that we're
removing that stale file which should alert the admin. In case of PID
re-use we have a new debug log message that could be used to diagnose
what's happening. In that case the admin could choose to determine why
this PID re-use is happening and how to potentially avoid it.

Closes #196.