cli: Make response inspection more robust
On server-side errors we'd see KeyError exceptions from the CLI because
'job_id' would not be present. The actually present 'message' field
would never reach the user. Add some simple logic so that we cope with a
non-existant job ID and give the user some feedback on it, including a
possibly present message from the server.