systemd: Streamline service unit
The PIDFile is not used by (or at the very least not useful for) systemd
because we're a service of type notify, i.e. are not forking into the
background and have a direct line of communication with systemd. So it
knowns our PID anyway and therefore is able to send us signals without a
PID file.

The KillMode defaults to control-group. So we can remove that setting.