config: Do not create PID file by default
Creating a PID file is redundant in systemd and container use-cases
which are the main ones nowadays. So we change the default to not
creating a PID file. Other use-cases such as traditional SysV-style init
can re-enable it using the peekaboo.conf option whose description is
updated to that effect.