ci: Add GitHub action for compose-based CI
Add an initial GitHub Actions workflow that checks out the
PeekabooAV-Installer and PeekabooAV repos in the correct structure to
then run docker-compose to bring up a CI pipeline for validation. A
simple test job is submittet that is supposed to always return verdict
'bad'. That way we test peekaboo-util, peekaboo proper and Cortex up to
and including its FileInfo analyser.