cortex: Update FileInfo analyzer version to 8.0
Cortex have updated their FileInfo analyzer to version 8.0 which changes
the name we need to use to access it. So update the version number. No
other changes seem necessary.

This could make us potentially backward-incompatible with older versions
of Cortex. But since the analyzers.json URL is not versioned, it appears
they're fully backward-compatible and easy to update (or even
auto-updated) Cortex-side. So we should address that concern only if it
becomes a problem.

A short review of Cortex's analyzers.json reveals no other version
changes.