expressions: Fix logic operators
Our implementation of logic operators 'and' and 'or' was wrong: They
used the bitwise operations from module 'operator'. This would work as
long as only booleans were used with them. But if anyone ever tried
something like 'True or 2' they'd produce '3' instead of 'True' because
'True' gets coerced into '1' instead of '2' getting coerced into 'True'.
We fix that by switching to small lambda functions that use the actual
python operators. Testsuite tests are added to document the problem and
catch regressions.