expressions: Ignore non-iterability of None in membership tests
Python makes a distinction between something being contained in
something else (or not) and something not being able to be contained in
nothing (i.e. None) and raises a 'NoneType not iterable' exception in
the latter case. We can see no application of this distinction in our
use-case as of now. In fact it causes problems when writing tests
against object properties which are None. Therefore we handle that case
in our containment tests and interpret it as something (obviously) not
being contained in nothing.

Closes #176.