cuckoo, cortex: Remove startup connectivity check
Checking connectivity at startup is not a good practice in highly
distributed systems like we're building usually: It needlessly creates
race-conditions during startup which need to be handled externally or by
exceedingly large timeouts and ultimately still do not work reliably. A
succeeding initial connectivity check still does not protect from the
resource becoming unavailable later on.

Instead we should handle errors as they occur during operation. We're
already doing that by failing job submission and retrying job tracking,
the latter even with a maximum job age.

So we remove the initial connectivity check from Cuckoo and Cortex
modules. This increases the likelihood of analyzer job submission
failure during system startup if Peekaboo is already up but Cuckoo or
Cortex aren't. It's the client's responsibility to resubmit its sample
in that case, as it would already need to do if the same occured during
normal operation e.g. due to an analyzer system maintenance or restart
or network outage.