cortex: Remove dependency on cortex4py
In the past we've repeatedly needed additional functionality in
cortex4py. With the new release looming and not all of our extensions
added upstream we can either push for inclusion, fork or replace it. All
options have problems. The latter however provides an opportunity to
evaluate how much functionality we need in the first place considering
that we might want to switch to asyncio and therefore aiohttp (instead
of requests) at some point.

It turns out that as of now we (similarly to Cuckoo) need only two
functions: Job submit and report retrieval. These can be implemented
using the by now well-understood requests mechanics from Cuckoo. We also
add schema validation right away.

Job submission has to be done in two steps: First looking up the ID of
the analyser by name and then submitting to that analyzer ID. Doing this
on each submission makes us cope with reconfiguration of Cortex at
runtime leading to changed analyser IDs.

As a change in behaviour, we now also have the option to include the
sample's type_declared in the job submission where cortex4py would use
the magic module to guess the type. It's unclear at this point what
difference it makes in practice but it certainly appears more consistent
that way.