Oletools: Remove warning
Our Oletool log a warning every time an empty file oder plain text file is
passed to olevba. This is because oletools assume as ultimate option the file to
analyse is the plain text vba code and hece won't fix.

We keep the test and remove the false result from the report but won't log any
more.