daemon: Gather up and exit on unexpected errors
Our running of workers, job trackers and the cluster duplicate handler
as asyncio tasks shows any unexpected error conditions by way of
exceptions only on shutdown. This was also the case with classic
threading and threads (and now tasks) would die silently, stop
functioning and even prevent regular shutdown.

To finally remedy this situation we make all startup routines that
create asyncio tasks return those tasks so we can build a list of
awaitables from which to gather any error conditions immediately when
they happen. A single call to asyncio.gather with these awaitables now
becomes the central serialisation point in the daemon instead of
PeekabooServer.serve().

The latter doesn't even make it into the list of awaitables since
asyncio servers in general and Sanic servers in particular run fully
enclosed in the event loop and do not expose any means to gather their
error conditions. Here we need to continue to rely on Sanic's built-in
error handling and logging (for now).

This restructuring now also allows us to fully start up the server
before reporting readiness to systemd.

Logging is restructured and extended to make every component report
startup and shutdown in a similar fashion.