config: Support drop files
Add functionality to load additional drop-in config snippets from a
directory ending in .d next to the main configuration file. This allows
to separate a distributed default config from local changes. Drop files
have to be named using leading double digits and a trailing .conf by
default. The former allows for lexicographical sorting determining the
order in which the snippets are read and option values overwrite each
other.

Two new global options allow customisation of the drop file directory
template and drop file globbing pattern. The test suite and main config
file sample are extended to match.

Because the logic is implemented in PeekabooConfigParser, it works with
all config files, not only the main peekaboo.conf.

Closes #69.