config: Move list correlation to an interpolation
We can use ConfigParser's interpolations to do correlation of multiple
options into lists as we need it for our rules and regular expression
lists. This somewhat simplifies the code. But more importantly it moves
correlation to read time which will allow us to implement additional
features such as resetting when reading from multiple files.