config: Allow for resetting lists
Drop files should be able to reset a list in order to easily override
its value in its entirety. So we add a special syntax where both
distinguisher and value are the dash (-). We implements it in our
Interpolation but need to do some additional post-processing in the
PeekabooConfigParser as well so the ConfigParser routines keep option
order from the drop file. The test suite is extended to match and a note
on the syntax is added to ruleset.conf.sample.