ci: Add scheduled build and push of container images
Regularly rebuilding all (supported) container images helps catch
regressions through external dependencies and keeps them up-to-date with
security patches. Therefore we add a scheduled Github Actions workflow
that does just that. It's using a dynamic matrix build that determines
the refs to build based on a filter list applied to the PeekabooAV repo
refs. It recycles all of the existing logic for image builds by moving
it into a custom composite action that is parametrised appropriately.

A composite action is used instead of a reuseable workflow because the
latter can not currently be used in conjunction with matrix builds.

The existing event-triggered workflow is switched to that action as
well.

Because we do not want to move the action to its own repo (for now), we
need to check out the PeekabooAV repo from the workflows before we're
able to use it. Later checkout actions insist on totally cleaning out
the workspace. Therefore we have to protect the action from deletion by
copying out of the workspace. Otherwise an implicit post step will fail
because it cannot find the action any more.

Unfortunately, the docker metadata action does not cater well to image
rebuilds. It basically assumes all releases to monotonically increase
versions and never being processed simulataneously or out of order. It
also assumes to be run on the source code corresponding to the
github.ref the workflow was called upon. This is only true for PRs,
pushes and taggings but not for scheduled builds.

Therefore we use a patched version of the action that accepts the ref to
work with as an input. Additionally we determine the latest version and
latest release per major release in preparation of the scheduled matrix
build and provide those the action where it influences the tagging
scheme.