knowntools: Improve database efficiency
Knowntools would pull a list of all previous analyses regarding a
particular sample from the database to determine the first, last and
worst analysis locally. The retrieval of that list would linearly scale
in time with the number of times a sample was submitted for analysis.
Since the journal was also kept in the report in full, it would also
linearly scale in memory consumption for the duration of the analysis.
Both is not ideal.

Therefore we switch to a more selective approach that requests only the
three data items of interest from the database in three distinct
queries. These are limited to a result count of one which, paired with
ordering by analysis time or result, extracts the analysis with the
required property of being oldest, most recent or worst.

The price to pay is twofold: For best efficiency with MariaDB/MySQL we
need two indices differing in the sorting column.

Also, since sqlite has no native support for enums, we can not use the
features provided by MariaDB/MySQL and PostgreSQL of sorting based on
declaration order of union members (which we could make sure to be
governed by our internal "worseness" metric). Therefore we add a
result_numeric column to table sample_info. We keep the result column
for easy use in queries with SQLAlchemy's enum support and human
readability.