knowntools: Improve database efficiency
Knowntools would pull a list of all previous analyses regarding a
particular sample from the database to determine the first, last and
worst analysis locally. The retrieval of that list would linearly scale
in time with the number of times a sample was submitted for analysis.
Since the journal was also kept in the report in full, it would also
linearly scale in memory consumption for the duration of the analysis.
Both is not ideal.

Therefore we switch to a more selective approach that requests only the
three data items of interest from the database in three distinct
queries. These are limited to a result count of one which, paired with
ordering by analysis time or result, extracts the analysis with the
required property of being oldest, most recent or worst.

The price to pay is twofold: For best efficiency with MariaDB/MySQL we
need two indices differing in the sorting column.

Also, since sqlite has no native support for enums, we can not use the
features provided by MariaDB/MySQL and PostgreSQL of sorting based on
declaration order of union members (which we could make sure to be
governed by our internal "worseness" metric). Therefore we add a case
statement to sort the result names by the numerical values from the
enum. This leads to an SQL statement like this:

SELECT sample_info_v10.analysis_time, sample_info_v10.result,
  sample_info_v10.reason FROM sample_info_v10
WHERE sample_info_v10.id != 20 AND
  sample_info_v10.result != 'failed' AND
  sample_info_v10.state = 'FINISHED' AND
  sample_info_v10.identity = '1518[...]9df8'
ORDER BY CASE sample_info_v10.result
  WHEN 'unchecked' THEN 1
  WHEN 'unknown' THEN 2
  WHEN 'ignored' THEN 3
  WHEN 'good' THEN 4
  WHEN 'failed' THEN 5
  WHEN 'bad' THEN 6
END DESC