Switch Known rule back from Knowntools
In case the ruleset does not use the Knowntools report through
expressions, the Known rule only needs the worst analysis result, not
the first and not the last. So we can in theory save another two (quite
complex and expensive) database requests per analysis if we switch the
Known rule back to directly accessing the database.

The drawback is that this result, unlike the Knowntools report, is not
cached in the sample. So if the sample does another (or more)
roundtrip(s) through the ruleset after some other analyser report became
available, the database will be queried again and again, negating the
initial savings.

The likelihood that the result of these queries will change and
short-circuit the analysis is quite low since we spend a lot of effort
on avoiding concurrent analyses of the same sample both locally and
cluster-wide.

But: If we actually want to pick up analysis results on the same sample
from other instances in a cluster we *need* the report not to be cached.
This was not a problem as long as duplicate samples were detected and
withheld before entering ruleset processing. But with an upcoming move
of duplicate detection into the ruleset this will no longer be the case.