Run testsuite on Ubuntu 20.04 for Python 3.6
Github Actions has moved ubuntu-latest to 22.04. This is fine since we
test with generic Python versions. But Python 3.6 is no longer provided
on 22.04. So we add an exception to the test matrix to run the testsuite
on Ubuntu 20.04 for Python 3.6. This is done by way of adding a new
matrix variable os which defines the runner image to use. This still
defaults to ubuntu-latest. An explicit matrix variable combination is
added with ubuntu-20.04 and python version 3.6.